requirements.txt
streamlit==1.49.1
python-docx==1.2.0
pandas==2.3.2
pillow==11.3.0
